podcheck.t: Try harder to avoid transitory failure
It may be that a file that looks like a pod is transitory.  podcheck.t
catches and handles some of these, but there are cases where it doesn't
catch these.  This commit adds code to try harder to detect such
instances and recover properly.

I know of two instances in the months since this code has been in
service where this has happened, so it's not a common occurrence, and
re-running it makes the failure go away.  Still, if it can be easily
avoided, do so.