[perl #111462] Don’t leak eval "" op tree when croaking
This patch only fixes the problem for croaks that occur in the peep-
hole optimiser or in Perl_finalize_optree.

It does this by doing SAVEFREEOP first and then restoring the
savestack index to its previous value afterwards (to void the effect
of SAVEFREEOP).

A more correct fix might be to do op_free in die_unwind before
POPEVAL, but I would have to do a lot more digging through the code
to tell whether that is safe.  I don’t feel comfortable with doing
that for 5.16.

This leak causes this warning on non-threaded debugging builds:

$ PERL_DESTRUCT_LEVEL=1 ./perl -Ilib -e 'BEGIN { $^H{foo} = bar } our %FIELDS; my main $x; eval q[$x->{foo}]'
Unbalanced string table refcount: (1) for "foo" during global destruction.

This problem does not affect the main program, because perl_destruct
frees PL_main_root.  It does not affect subroutines, as the op tree is
attached to the CV first, so freeing the CV frees the op tree.