[perl #111864] Don’t leave obj on stack for -x $overloaded
Commit 8db8f6b697e changed the way filetest operators use the stack,
in order to make stacked -t work.

Filetest operators were changed to pop the argument off the stack for
a standalone operator, but to leave it on the stack for a stacking op
(so the next operator could use it).  The code for handling overloaded
objects, which was separate, was already doing something similar, by
not popping the object off the stack.

I made the mistake of changing overloaded objects’ return code to
share code with regular filetest operators (the FT_RETURN_* macros),
but without changing the way the overload code got the item from the
stack to begin with.  Hence, the returning code assumed that the argu-
ment had been popped for non-stacking ops, which was not the case.

This commit changes the way the overload case does it’s return, to
account for the fact that the object must be left on the stack when
initially fetched from it (in case the object turns out not to have -X
overloading and the regular code has to kick in).