Remove 2 microperl suggestions from the todo list.
The --regen option added to Porting/checkcfgvar.pl makes creating uconfig.sh
pretty much automatic.

Whilst maintaining Makefile.micro could be made more automatic by templating
it based on MANIFEST, the benefits are probably never going to pay back the
cost of doing it, as we add (or remove) C and header files so rarely.
It would only really make sense as part of a system that did the same for
Makefile.SH, vms/descrip_mms.template and the multitude of Win* Makefiles.