Safe.pm: Don’t eval code under ‘no strict’
Instead of evaluating code under ‘no strict’, we should be evaluating
it with no pragmata at all by default.

This allows ‘use 5.012’ to enable strictures in reval.  It also
has the side effect of suppressing the ‘Unbalanced string table
refcount’ warnings, at least in some cases.  This was brought up in
ticket #107000.