[perl #112352] Make ‘use v5.15’ always enable unicode_eval
When the unicode_eval feature was added, the code for compiling an
eval was modified to check whether the feature was enabled, before
flagging the op.

For efficiency’s sake, since there was already a block that checked to
see whether %^H was localised (which happens automatically when it is
modified), and since features always used %^H, I put the check inside
that block, so it would be skipped altogether without %^H set.  (It’s
debatable whether it was actually any faster.)

Later, in commit 6634bb9d0ed1, when I changed implicit features to use
$^H, I didn’t update the eval code.  So it would only check to see
whether the feature was enabled if %^H was also enabled.

This fixes that by moving it below the %^H block.