In pp_pow, avoid special case code for an IV_MIN result.
Given that IV_MIN on a twos complement system will always be a (negated) power
of two, and powers of two are handled in the NV-based code above, a IV_MIN
result can't actually happen. However, by switching the integer result code
to the same structure as the basic arithmetic ops, we can handle it as part of
the code dealing with regular negative results. Making similar code similar
eases maintenance.