In pp_pow, remove calls to SvIV_please_nomg() on (not the) result.
Commit 52a96ae66a5b0cd1 in Feb 2003 added logic to call SvIV_please(TOPs)
for the places where pp_pow had integer inputs, but returned an NV. TOPs
was used, because at the point of return, the result is in the top slot of
the stack. Commit 800401ee2a8a5a67 in Oct 2007 inadvertently changed the TOPs
to svr, as part of a larger (systematic) refactoring that needed to use local
variables instead of stack positions to reference the input SVs, not
realising that these two particular instances of TOPs were not referencing
the right input operand. No tests failed.

(The SvIV_please() was subsequently changed to *_nomg() as part of commit
6f1401dc2acd2a2b)

Hence the code as-is is wrong. However, even changing these two back to TOPs
to massage the result SV is unnecessary, as they will have no Perl-space
effect. If the NV value is not in the range that integers preserve UVs,
then the public IOK flag will not be set. If it is, that flag will be set,
but all arithmetic operations *also* call SvIV_please_nomg() on their input,
so the flag would be set there (instead).

So remove the code, as it's not needed.