Remove dead MSDOS code from S_init_postdump_symbols().
The original MSDOS port of 3.0 and 4.0 was removed in 5.000, but its #ifdefs
remained in the code. The djgpp port was added post 5.004, and djgpp
defines both MSDOS and DJGPP. Hence this code, which was made conditional on
MSDOS but *not* DJGPP by commit 91a64263ab3d9ea5 in Aug 2003 should instead
have been removed, as it's dead.