Various minor fixes to enable -DUSE_SFIO to compile again.
Historically, stdio was "enabled" with -UUSE_PERLIO, PerlIO (proper) enabled
with -DUSE_PERLIO, and sfio enabled with -DUSE_PERLIO && -DUSE_SFIO.

Commit 39f7a87036eb8d13 in Apr 2003 introduced PerlIO::get_layers, but
broke the build for stdio & sfio. Commit 5fef3b4a7f0c97d3, the next day,
restored the build for stdio, but not for sfio. This is fixed with a 1 line
change in universal.c

Commit de009b76d60bdeb8 in Apr 2005 erroneously adds const to a variable
which is not constant in PerlIO_findFILE(). As this was in code only
conditionally compiled with -DUSE_SFIO it was not noticed.

Commit 6cb8cb214e85c7d6 in Oct 2006 added PerlIO_teardown, which breaks the
sfio build. A 1 line change in perl.h fixes this.

Commit 4ee3916907333148 in Nov 2008 made a typo in code conditionally
compiled with -DUSE_SFIO within Perl_write_to_stderr().

Commit 2e0cfa16dea85dd3 in Feb 2011 broke the build for stdio & sfio.
Commit b6ae43b72863e12d in Mar 2011 restored the build for stdio, but not
for sfio. This is fixed with a 1 line change in util.c

Even with these changes, miniperl is unable to build all extensions in
ext, dist and cpan, due to an undiagnosed bug which causes generated
Makefiles to be truncated at 8192 bytes. Additionally, miniperl built with
sfio fails many many tests for make minitest, and hangs for input on
t/op/readline.t