[perl #111794] Make goto "" like goto ${\""}
The logic was written in such a way that goto "" just happened to slip
past all the checks and cause pp_goto to return NULL for the next op,
which means the end of the program.

goto ${\""} dies with ‘goto must have label’, so goto ""
should as well.

This also adds other tests for that error, which was apparently
untested till now.