Remove todo for UTF8 source filters
Source filters don’t really make sense on character streams.  They are
designed for streams of bytes coming straight from a file.  Things
stop making sense if you have ‘use utf8’ in a UTF-8 scalar (does that
mean double-decode?).

It’s for this reason that evalbytes respects source filters, while
eval does not.  (It doesn’t outside the unicode_eval feature, because
it was never really thought about and the implementation didn’t take
it into account, resulting in strange behaviour.  It doesn’t with the
unicode_eval feature, because it was intentionally prohibited.)