No need to prefix the $(CC) invocation that links miniperl with $(LDLIBPTH).
$(LDLIBPTH) is used to prefix the appropriate LD_LIBRARY_PATH=... before
commands in the Makefile when perl is build with a shared perl library so
that the uninstalled ./perl will be able to find it.

Commit c4f23d77f4b3486a (in May 1998) added the code for LDLIBPTH, but
also added it to the default command line used to invoke $(CC) to link
miniperl. This command doesn't need to run ./perl, hence the use of
$(LDLIBPTH) is superfluous here. Removing it makes the code simpler.