Fix a dependency issue when building a shared libperl on AIX.
AIX requires a list of exported symbols in the file perl.exp, which is
generated by makedef.pl. For the non-shared libperl case, the macro
PERLEXPORT is set to this, and $(PERLEXPORT) is a pre-requisite for building
dynamic extensions. However, historically, when building a shared libperl,
Makefile.SH was not setting PERLEXPORT, and instead setting LIBPERLEXPORT.
However, there was no dependency on this macro for building dynamic
extensions. Hence the build can fail if make chooses to build dynamic
extensions before perl.exp

The simplest fix seems to be to use the same name, PERLEXPORT, in both
cases. PERLEXPORT isn't referenced anywhere in the Makefile other than as
pre-requisite for dynamic extensions and the main perl binary.

The use of LIBPERLEXPORT dates back to commit 549a6b102c2ac8c4 in Jul 1999,
"Fixed AIX dynamic loading and AIX shared Perl library." It's surprising
that this problem hasn't bitten anyone previously.