File::Glob: Don’t use the magic 2nd arg to glob
This argument is going away, because it is undocumented, unused on
CPAN outside of the core, and getting in the way of allowing glob() to
be overridden properly.

It gets in the way because, under the ‘overrides’ feature (still a
no-op), the glob keyword will be truly overridable, with no magic
involved, but it will still be called by the built-in <...> operator,
which consequently must not pass any magic second argument.

See <https://rt.perl.org/rt3/Ticket/Display.html?id=108286>.