Eliminate the special case linking of miniperl on darwin, which is redundant.
Commit cb3fc4263509f28c (May 2003) removed the use of -flat_namespace from
the link flags, but added it specially in Makefile.SH for miniperl, so that
symbols from opmini.o overrode those in libperl.dylib. However, a side effect
of commit 908fcb8bef8cbab8 (Dec 2006) was to change the linker line to use
explicit object files, meaning that op.o was no longer part of linking,
meaning that the override is no longer needed. Hence darwin's link does not
need special-casing.