Only set CCDLFLAGS once in the Makefile on AIX.
Commit 5f9d9a1727238445 (Jul 1999) fixed Makefile.SH so that CCDLFLAGS
correctly pointed the compiler to the uninstalled perl.exp, instead of the
config.sh value which reflected the (not yet) installed location. However,
the commit assumed that it needed to redefine CCDLFLAGS part way through
the Makefile. This is unnecessary, as the macro CCDLFLAGS is not used for
any purpose other than the local build. So refactor Makefile.SH to generate
a Makefile which only sets CCDLFLAGS once.