Make ‘require 6’ a syntactic special case
This commit removes the special-casing for ‘require 6’ from the token-
iser and from pp_require.  Now it all happens at optree-building time
in op.c.  If the argument to require is a constant (not counting
folded constants) that is a number or a vstring, the require
op is flagged as being a version check, with the newly-added
OPpREQUIRE_VER flag.

This means that ‘require $ver’ no longer works, which could be consid-
ered an incompatible change (some tests had to change), but see the
bug fixes below.

Since CORE::require($_[0]) will no longer work from an override (actu-
ally a callback; see perl #108286), overrides are no longer called for
version checks like ‘require 5’.  The tests for this are marked to-do
for now, as they will be re-enabled (with ‘use feature "overrides"’)
when the ‘overrides’ feature comes into effect.

These bug fixes result from this change:

• [perl #105926] These two lines:

    $x = "6"; require $x;
    $x = "6"; 0+$x; require $x;

  should behave exactly the same way, but the latter was a version
  check.  Now they are both file names. And this no longer happens:

    $ perl5.12.4 -le '$x = "34 cabbages"; 0+$x; require $x'
    Invalid version format (non-numeric data) at -e line 1.

  In this case, "34 cabbages" should obviously be a file name, but its
  numeric representation was causing it to be treated as a version
  number; yet the version-parsing code used the string representation.

• [perl #105924] An infix operator that has a different meaning when
  an infix operator is not expected (e.g., <<) is now permitted after
  a number immediately following ‘require’, just like infix operators
  that are always unambiguous (such as >>).  The second of these two
  lines used to be a syntax error:

    print require 5 >> $3;
    print require 5 << $3;

• [perl #105928] require no longer reports a syntax error when certain
  characters follow a version number.  Previously, only ; { } and
  whitespace were allowed after it.  The first of these lines used to
  be a syntax error, while the second was fine:

    require 5, print "ok";
    require 5 , print "ok";

Those last two bug fixes are due to the removal of special-casing from
the tokeniser.  ‘require’ was using the same version-parsing routine
as ‘use’, which seems to have been written for the sake of ‘use’
(which has more stringent parsing requirements, being forbidden in an
expression), without taking ‘require’ into account.